+++ src/commands/deploy.js
@@ -3,7 +3,7 @@ const { Command, flags } = require('@oclif/command')
 const { output } = require('../services/flags')
 
 const { printDeployResults } = require('../services/utils')
-const { runDeploy } = require('../services/deploy')
+const { runDeploy } = require('../modules/deploy')
 
 class DeployCommand extends Command {
   async run() {
