+++ src/modules/deploy/index.js
@@ -1,0 +1,5 @@
+const runDeploy = require('./deploy')
+
+module.exports = {
+  runDeploy,
+}
